kernel: add permutat_intern.hh
