+++ src/pperm.cc
@@ -42,6 +42,8 @@ extern "C" {
 
 } // extern "C"
 
+#include "permutat_intern.hh"
+
 
 //
 // convert TNUM to underlying C data type
@@ -104,66 +106,6 @@ static inline UInt DEG_PPERM(Obj f)
     ASSERT_IS_PPERM<T>(f);
     return (UInt)(SIZE_OBJ(f) - sizeof(T) - 2 * sizeof(Obj)) / sizeof(T);
 }
-
-
-//
-// Various helper functions for permutations
-//
-template <typename T>
-struct T_PERM {
-};
-template <>
-struct T_PERM<UInt2> {
-    static const UInt tnum = T_PERM2;
-};
-template <>
-struct T_PERM<UInt4> {
-    static const UInt tnum = T_PERM4;
-};
-
-template <typename T>
-static void ASSERT_IS_PERM(Obj perm)
-{
-    GAP_ASSERT(TNUM_OBJ(perm) == T_PERM<T>::tnum);
-}
-
-template <typename T>
-static inline UInt DEG_PERM(Obj perm)
-{
-    ASSERT_IS_PERM<T>(perm);
-    return (SIZE_OBJ(perm) - sizeof(Obj)) / sizeof(T);
-}
-
-template <typename T>
-static inline T * ADDR_PERM(Obj perm)
-{
-    ASSERT_IS_PERM<T>(perm);
-    return (T *)(ADDR_OBJ(perm) + 1);
-}
-
-template <typename T>
-static inline const T * CONST_ADDR_PERM(Obj perm)
-{
-    ASSERT_IS_PERM<T>(perm);
-    return (const T *)(CONST_ADDR_OBJ(perm) + 1);
-}
-
-
-//
-// The 'ResultType' template is used by functions which take two partial
-// permutations as argument to select the type of the output they produce: by
-// default, a T_PPERM4, whose entries are stored as UInt4. But if both inputs
-// are T_PPERM2, then as a special case the output is also a T_PPERM2, whose
-// entries are stored as UInt2.
-//
-template <typename TL, typename TR>
-struct ResultType {
-    typedef UInt4 type;
-};
-template <>
-struct ResultType<UInt2, UInt2> {
-    typedef UInt2 type;
-};
 
 
 #define MAX(a, b) (a < b ? b : a)
