+++ src/trans.cc
@@ -66,6 +66,8 @@ extern "C" {
 
 } // extern "C"
 
+#include "permutat_intern.hh"
+
 
 #define MIN(a, b) (a < b ? a : b)
 #define MAX(a, b) (a < b ? b : a)
